+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.1] - 2025-08-10
+
+### Changed
+
+- Tightened the argument validation and reduce runtime error paths in [main.rs](src/main.rs)
+- The app now warns the user when the window title cannot be set in [app.rs](src/ui/app.rs)
+- Avoid panicking when the scroll position is out of bounds in [app.rs](src/ui/app.rs)
+- Use constants for UI elements in [app.rs](src/ui/app.rs)
+
 ## [0.5.0] - 2025-08-09
 
 ### Changed
@@ -17,7 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove unnecessary `std::sync::Mutex` from `LOG_FILE_PATH` as `std::sync::LazyLock` is
 thread-safe in [logging.rs](src/ui/logging.rs)
 
-## [0.4.3] -- 2025-08-07
+## [0.4.3] - 2025-08-07
 
 ### Changed